Compiled with -Werror and fixed ALL the warnings. Gah, about time!
git-svn-id: http://quadra.googlecode.com/svn/trunk/quadra@498 8926ee93-8b47-0410-9975-e57e0fb0bd0c

+++ skelton/common/net.cpp
@@ -1017,7 +1017,7 @@ void Net::sendudp(Dword to, Packet *p) {
 		int temp = sendto(udpsock[i], (const char *) nb.buf, nb.len(), 0, (sockaddr *) &udpsin, sizeof(udpsin));
 		if(getlasterror(temp))
 			return; // failed!
-		if(temp != nb.len()) {
+		if(temp != static_cast<int>(nb.len())) {
 			last_error = "Protocol UDP did not send all requested data";
 			return; // failed!
 		}
